+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## [0.7.2] - Unreleased
+## [0.7.3] - Unreleased
+### Added
+* n/a
+### Fixed
+* n/a
+### Updated
+* n/a
+### Breaking changes
+* n/a
+
+## [0.7.2] - 2025-04-03
 ### Added
 * Lookup warehouses by `name` for Terraform backend
 * Databricks `Permissions` declaration from the stack root.